+++ apps/web/src/components/preview/chat/ChatPreview.tsx
@@ -4,12 +4,12 @@ import { useAuthController } from '@novu/shared-web';
 import { useMutation } from '@tanstack/react-query';
 import { useEffect, useState } from 'react';
 import { useFormContext, useWatch } from 'react-hook-form';
-import { previewChat } from '../../../../api/content-templates';
-import { getLocalesFromContent } from '../../../../api/translations';
-import { IForm } from '../../../../pages/templates/components/formTypes';
-import { useStepFormCombinedErrors } from '../../../../pages/templates/hooks/useStepFormCombinedErrors';
-import { useStepFormPath } from '../../../../pages/templates/hooks/useStepFormPath';
-import { errorMessage } from '../../../../utils/notifications';
+import { previewChat } from '../../../api/content-templates';
+import { getLocalesFromContent } from '../../../api/translations';
+import { IForm } from '../../../pages/templates/components/formTypes';
+import { useStepFormCombinedErrors } from '../../../pages/templates/hooks/useStepFormCombinedErrors';
+import { useStepFormPath } from '../../../pages/templates/hooks/useStepFormPath';
+import { errorMessage } from '../../../utils/notifications';
 import { LocaleSelect } from '../common';
 import { ChatContent } from './ChatContent';
 import { ChatInput } from './ChatInput';
