+++ apps/web/src/components/preview/common/EmojiIcon.tsx
