+++ apps/web/src/components/preview/common/icons/android/AndroidIcon.tsx
