+++ apps/web/src/components/preview/common/icons/android/AndroidIndicatorsIcon.tsx
