+++ apps/web/src/components/preview/common/icons/iphone/AppleIcon.tsx
