+++ apps/web/src/components/preview/common/icons/iphone/IOSIndicatorsIcon.tsx
