+++ apps/web/src/components/preview/email/PreviewSegment/MobileIcon.tsx
