+++ apps/web/src/components/preview/email/PreviewSegment/WebIcon.tsx
