+++ apps/web/src/components/preview/email/PreviewWeb.tsx
@@ -4,9 +4,9 @@ import { useState } from 'react';
 import { ErrorBoundary } from 'react-error-boundary';
 import Frame from 'react-frame-component';
 import { FieldError, FieldErrorsImpl, Merge } from 'react-hook-form';
-import { IFormStep } from '../../../../pages/templates/components/formTypes';
-import { EmailIntegrationInfo } from '../../../../pages/templates/editor/EmailIntegrationInfo';
-import { When } from '../../../utils/When';
+import { IFormStep } from '../../../pages/templates/components/formTypes';
+import { EmailIntegrationInfo } from '../../../pages/templates/editor/EmailIntegrationInfo';
+import { When } from '../../utils/When';
 import { LocaleSelect } from '../common/LocaleSelect';
 import { PreviewEditOverlay } from '../common/PreviewEditOverlay';
 import { PreviewUserIcon } from '../common/PreviewUserIcon';
