+++ apps/web/src/components/preview/email/Skeleton/ContentSkeleton.tsx
