+++ apps/web/src/pages/templates/components/email-editor/EmailMessagesCards.tsx
@@ -2,13 +2,13 @@ import { useState } from 'react';
 import { EmailContentCard } from './EmailContentCard';
 import { useAuthContext } from '../../../../components/providers/AuthProvider';
 import { When } from '../../../../components/utils/When';
-import { EmailPreview } from '../../../../components/workflow/preview';
+import { EmailPreview } from '../../../../components/preview';
 import { EditorPreviewSwitch } from '../EditorPreviewSwitch';
 import { Grid, SegmentedControl, useMantineTheme } from '@mantine/core';
 import { TestSendEmail } from './TestSendEmail';
 import { colors } from '@novu/design-system';
-import { MobileIcon } from '../../../../components/workflow/preview/email/PreviewSegment/MobileIcon';
-import { WebIcon } from '../../../../components/workflow/preview/email/PreviewSegment/WebIcon';
+import { MobileIcon } from '../../../../components/preview/email/PreviewSegment/MobileIcon';
+import { WebIcon } from '../../../../components/preview/email/PreviewSegment/WebIcon';
 import { useHotkeys } from '@mantine/hooks';
 import { VariablesManagement } from './variables-management/VariablesManagement';
 import {
